refactor(common): move assert tools to new `test::assert` module
